--- lib/data/api/api_constants.dart
@@ -1,10 +1,0 @@
-class ApiPath {
-  ApiPath._();
-  static const users = 'users/';
-}
-
-class BaseUrl {
-  BaseUrl._();
-  static const baseUrlDev = 'https://api.github.com/';
-  static const appApiKey = 'api key TODO';
-}
