+++ lib/data/api/cient/api_client.dart
@@ -1,9 +1,10 @@
 import 'package:dio/dio.dart';
-import 'package:flutter_riverpod_template/data/api/api_constants.dart';
+import 'package:flutter_riverpod_template/data/api/cient/api_constants.dart';
 import 'package:flutter_riverpod_template/feature/repository_list/models/repository_list_model.dart';
 import 'package:retrofit/retrofit.dart';
 
 part 'api_client.g.dart';
+
 
 @RestApi()
 abstract class ApiClient {
