+++ lib/data/api/cient/api_constants.dart
@@ -1,0 +1,4 @@
+class ApiPath {
+  ApiPath._();
+  static const users = 'users/';
+}
