+++ lib/data/api/providers/user/user_repository.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_riverpod_template/data/api/api_client.dart';
+import 'package:flutter_riverpod_template/data/api/cient/api_client.dart';
 import 'package:flutter_riverpod_template/feature/repository_list/models/repository_list_model.dart';
 
 class UserRepository {
